Add backup logic

+++ backuptool/backup.py
@@ -13,8 +13,15 @@ def run_backup(
         target_dir: str | PathLike[str] = Path.cwd(),
         ignore: list | None = None,
         name_format: str = "%Y%m%d_%H%M%S", # strftime format string
-        archive_type: Literal["zip", "tar", "gztar", "bztar", "xztar"] = "zip",
+        archive_type: Literal["zip", "tar", "gztar", "bztar", "xztar"] | None = "zip",
 ):
+    """ Create backup of your project
+        :param root_dir: Absolute path to your project or project root folder name
+        :param target_dir: target folder for your backups
+        :param ignore: list of your files or folders you don't want to have in backup
+        :param name_format: strftime format string, default is ISO format
+        :param archive_type: Which type of archive you want, if you want to just copy your project without using archive use 'None'
+        :return Path to your backup"""
     ignore = ignore or []
     now = datetime.now()
     form = now.strftime(name_format)
@@ -29,8 +36,9 @@ def run_backup(
         func(path)
 
     shutil.copytree(root_dir, target_dir, ignore=ignore_files, dirs_exist_ok=True)
-    shutil.make_archive(str(target_dir), archive_type, target_dir)
-    os.chmod(target_dir, stat.S_IWRITE)
-    shutil.rmtree(target_dir, onerror=remove_readonly)
 
-    return archive_type
+    if archive_type is not None:
+        shutil.make_archive(str(target_dir), archive_type, target_dir)
+        shutil.rmtree(target_dir, onerror=remove_readonly)
+
+    return str(target_dir)
